+++ src/main/java/racingcar/dao/InsertCarDao.java
@@ -1,10 +1,10 @@
-package racingcar.repository;
+package racingcar.dao;
 
 import java.util.List;
 import javax.sql.DataSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import racingcar.repository.entity.CarEntity;
+import racingcar.dao.entity.CarEntity;
 
 public class InsertCarDao {
