+++ src/main/java/racingcar/dao/SelectWinnerDao.java
@@ -1,9 +1,9 @@
-package racingcar.repository;
+package racingcar.dao;
 
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
-import racingcar.repository.entity.WinnerEntity;
+import racingcar.dao.entity.WinnerEntity;
 
 public class SelectWinnerDao {
