+++ src/test/java/racingcar/repository/CarDaoTest.java
@@ -11,9 +11,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import racingcar.dao.CarDao;
 import racingcar.domain.RacingGame;
-import racingcar.repository.entity.CarEntity;
-import racingcar.repository.entity.GameEntity;
+import racingcar.dao.entity.CarEntity;
+import racingcar.dao.entity.GameEntity;
 
 @JdbcTest
 @SuppressWarnings("NonAsciiCharacters")
