+++ src/test/java/racingcar/repository/InsertRacingGameDaoTest.java
@@ -9,8 +9,9 @@ import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import racingcar.dao.InsertGameDao;
 import racingcar.domain.RacingGame;
-import racingcar.repository.entity.GameEntity;
+import racingcar.dao.entity.GameEntity;
 
 @JdbcTest
 @SuppressWarnings("NonAsciiCharacters")
