+++ src/test/java/racingcar/repository/WinnerDaoTest.java
@@ -11,10 +11,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import racingcar.dao.WinnerDao;
 import racingcar.domain.RacingGame;
-import racingcar.repository.entity.CarEntity;
-import racingcar.repository.entity.GameEntity;
-import racingcar.repository.entity.WinnerEntity;
+import racingcar.dao.entity.CarEntity;
+import racingcar.dao.entity.GameEntity;
+import racingcar.dao.entity.WinnerEntity;
 
 @JdbcTest
 @SuppressWarnings("NonAsciiCharacters")
