+++ src/test/java/racingcar/service/StubRacingGameRepository.java
@@ -1,7 +1,8 @@
 package racingcar.service;
 
 import racingcar.domain.RacingGame;
-import racingcar.repository.entity.GameEntity;
+import racingcar.repository.RacingGameRepository;
+import racingcar.dao.entity.GameEntity;
 
 public class StubRacingGameRepository implements RacingGameRepository {
